feat(ui): relocate tRPC MAO components to @nous/ui, wire MaoServicesProvider — WR-082 SP 1.2
- Move 4 tRPC-dependent components to @nous/ui with context injection
- Convert 3 test files to MaoServicesProvider-based mocking
- Rewire web app MAO page to import from @nous/ui
- Delete original self/apps/web/components/mao/ directory
- Fix MaoControlAuditHistory type mismatch (object→array)
- All 8 MAO components now in @nous/ui, zero tRPC/Next.js deps
- 3,046 tests passing, build clean
